requirements.txt
python-osrsapi==0.0.3